Move all logging items to separate file
