+++ go/pkg/utils/cliUtils.go
@@ -4,10 +4,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"os"
-
-	"github.com/go-logr/logr"
-	"go.uber.org/zap/zapcore"
-	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 )
 
 func PrintNoArgsWarningMessage(command string) {
@@ -50,33 +46,4 @@ func RedirectErrorToStdErrAndExit(err error) {
 func RedirectErrorStringToStdErrAndExit(err string) {
 	fmt.Fprintln(os.Stderr, err)
 	os.Exit(1)
-}
-
-var Logger logr.Logger
-
-func getZapcoreLevel(level string) (zapcore.Level, error) {
-	switch level {
-	case "debug":
-		return zapcore.DebugLevel, nil
-	case "warning":
-		return zapcore.WarnLevel, nil
-	case "info":
-		return zapcore.InfoLevel, nil
-	case "":
-		return zapcore.ErrorLevel, nil
-	default:
-		return nil, fmt.Errorf("log leve %s does not exist", level)
-	}
-}
-
-func GenLogger(logLevel string) error {
-	level, err := getZapcoreLevel(logLevel)
-	if err != nil {
-		return err
-	}
-	Logger = zap.New(zap.UseFlagOptions(&zap.Options{
-		Development: true,
-		TimeEncoder: zapcore.ISO8601TimeEncoder,
-		Level:       level,
-	}))
 }
